services/user_service/requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
passlib>=1.7.4
psycopg2-binary
confluent-kafka